styles/ subfolders. Remove enums folder. Modal css refactor.
