+++ client/src/components/modals/AreYouSureModal.js
@@ -2,11 +2,12 @@ import React, { useRef } from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
 
-import '../../styles/AreYouSureModal.css'
+import '../../styles/modals/AreYouSureModal.css'
 
+import Modal from './Modal'
 import { confirmAreYouSure, hideModal } from '../../actions/modalActions'
 
-function AreYouSureModal(props) {
+const AreYouSureModal = (props) => {
   const formRef = useRef(null)
 
   const handleFormSubmit = (e) => {
@@ -23,29 +24,27 @@ function AreYouSureModal(props) {
   }
 
   return (
-    <div class='modal-bg'>
-      <div class='are-you-sure-modal'>
-        <div>Are you sure?</div>
-        <form
-          id='areYouSureForm'
-          name='areYouSureForm'
-          onSubmit={handleFormSubmit}
-          ref={formRef}
+    <Modal class='are-you-sure-modal'>
+      <div>Are you sure?</div>
+      <form
+        id='areYouSureForm'
+        name='areYouSureForm'
+        onSubmit={handleFormSubmit}
+        ref={formRef}
+      >
+        <button
+          type='button'
+          onClick={handleClose}
+          id='areYouSureCancel'
+          class='btn'
         >
-          <button
-            type='button'
-            onClick={handleClose}
-            id='areYouSureCancel'
-            class='btn'
-          >
-            Cancel
-          </button>
-          <button type='submit' id='areYouSureSubmit' class='btn'>
-            Confirm
-          </button>
-        </form>
-      </div>
-    </div>
+          Cancel
+        </button>
+        <button type='submit' id='areYouSureSubmit' class='btn'>
+          Confirm
+        </button>
+      </form>
+    </Modal>
   )
 }
