+++ client/src/components/modals/Modal.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import '../../styles/modals/Modal.css'
+
+const Modal = (props) => {
+  return (
+    <div class='modal-bg'>
+      <div class={`modal ${props.class}`}>{props.children}</div>
+    </div>
+  )
+}
+
+export default Modal
