+++ client/src/components/recipe/Recipe.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 import { useCookies } from 'react-cookie'
 
-import '../../styles/Recipe.css'
-import '../../styles/Print.css'
+import '../../styles/recipe/Recipe.css'
+import '../../styles/recipe/Print.css'
 import Badge from '../widgets/Badge'
 import Nutrition from './Nutrition'
 import RecipeMadeButton from './RecipeMadeButton'
