+++ client/src/components/upload/EditForm.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import RecipePicker from './RecipePicker'
 import UploadForm from './UploadForm'
-import '../../styles/UploadForm.css'
+import '../../styles/upload/UploadForm.css'
 
 function EditForm(props) {
   const [recipeName, setRecipeName] = useState('')
