+++ client/src/components/widgets/Badge.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import '../../styles/Badge.css'
+import '../../styles/widgets/Badge.css'
 
 function Badge(props) {
   return <div class='badge badge-primary'>{props.children}</div>
