+++ client/src/styles/modals/AreYouSureModal.css
@@ -1,19 +1,7 @@
-.are-you-sure-modal {
-  position: absolute;
+.modal.are-you-sure-modal {
   width: 400px;
   max-width: 80%;
   height: 200px;
-
-  top: 50%;
-  transform: translateY(-50%);
-
-  padding: 30px;
-  border: 1px solid var(--dark-theme-form-control-border-color, black);
-  border-radius: 20px;
-
-  background-color: var(--dark-theme-bg, white);
-  color: var(--dark-theme-navbar-text, black);
-
   font-size: 24px;
   text-align: center;
 }
