+++ client/src/styles/modals/Modal.css
@@ -9,7 +9,7 @@
   background-color: rgba(0, 0, 0, 0.7);
 }
 
-.sign-in-modal {
+.modal {
   position: absolute;
   width: 400px;
   height: 500px;
