+++ client/src/styles/modals/SignInModal.css
@@ -1,0 +1,2 @@
+.modal.sign-in-modal {
+}
